Backend/pyproject.toml
[project]
name = "quid-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Mahima0Verma",email = "mahimavermakumari@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-filter (>=24.3,<25.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
